home_site/requirements.txt
django==4.2.2
djangorestframework==3.14.0
mysqlclient==2.2.4
tinytuya==1.12.8
django-cors-headers==4.3.1
